refactor(tree): Reorg modular change family format and codecs (#16013)
