+++ experimental/dds/tree2/src/feature-libraries/modular-schema/modularChangeCodecs.ts
@@ -3,28 +3,18 @@
  * Licensed under the MIT License.
  */
 
-import { Static, TAnySchema, Type } from "@sinclair/typebox";
+import { TAnySchema } from "@sinclair/typebox";
 import { assert } from "@fluidframework/common-utils";
 import {
 	FieldKey,
 	FieldKindIdentifier,
-	FieldKindIdentifierSchema,
 	GlobalFieldKey,
-	GlobalFieldKeySchema,
 	isGlobalFieldKey,
 	keyFromSymbol,
 	LocalFieldKey,
-	LocalFieldKeySchema,
-	RevisionTagSchema,
 	symbolFromKey,
 } from "../../core";
-import {
-	brand,
-	fail,
-	JsonCompatibleReadOnly,
-	JsonCompatibleReadOnlySchema,
-	Mutable,
-} from "../../util";
+import { brand, fail, JsonCompatibleReadOnly, Mutable } from "../../util";
 import {
 	ICodecFamily,
 	ICodecOptions,
@@ -33,94 +23,21 @@ import {
 	makeCodecFamily,
 	SchemaValidationFunction,
 } from "../../codec";
-import { ChangesetLocalIdSchema } from "./crossFieldQueries";
 import {
 	FieldChangeMap,
 	FieldChangeset,
 	ModularChangeset,
 	NodeChangeset,
 	RevisionInfo,
-} from "./fieldChangeHandler";
+} from "./modularChangeTypes";
 import { FieldKind } from "./fieldKind";
 import { genericFieldKind } from "./genericFieldKind";
-
-const EncodedValueChange = Type.Object(
-	{
-		revision: Type.Optional(RevisionTagSchema),
-		value: Type.Optional(JsonCompatibleReadOnlySchema),
-	},
-	{ additionalProperties: false },
-);
-type EncodedValueChange = Static<typeof EncodedValueChange>;
-
-const EncodedValueConstraint = Type.Object(
-	{
-		value: Type.Optional(JsonCompatibleReadOnlySchema),
-		violated: Type.Boolean(),
-	},
-	{ additionalProperties: false },
-);
-type EncodedValueConstraint = Static<typeof EncodedValueConstraint>;
-
-const EncodedFieldChange = Type.Object({
-	fieldKey: Type.Union([LocalFieldKeySchema, GlobalFieldKeySchema]),
-	keyIsGlobal: Type.Boolean(),
-	fieldKind: FieldKindIdentifierSchema,
-	// Implementation note: node and field change encoding is mutually recursive.
-	// This field marks a boundary in that recursion to avoid constructing excessively complex
-	// recursive types. Encoded changes are validated at this boundary at runtime--see logic
-	// later in this file's codec.
-	change: JsonCompatibleReadOnlySchema,
-});
-
-interface EncodedFieldChange extends Static<typeof EncodedFieldChange> {
-	/**
-	 * Encoded in format selected by `fieldKind`
-	 */
-	change: JsonCompatibleReadOnly;
-}
-
-const EncodedFieldChangeMap = Type.Array(EncodedFieldChange);
-
-/**
- * Format for encoding as json.
- *
- * This chooses to use lists of named objects instead of maps:
- * this choice is somewhat arbitrary, but avoids user data being used as object keys,
- * which can sometimes be an issue (for example handling that for "__proto__" can require care).
- * It also allows dealing with global vs local field key disambiguation via a flag on the field.
- */
-type EncodedFieldChangeMap = Static<typeof EncodedFieldChangeMap>;
-
-const EncodedNodeExistsConstraint = Type.Object({
-	violated: Type.Boolean(),
-});
-type EncodedNodeExistsConstraint = Static<typeof EncodedNodeExistsConstraint>;
-
-const EncodedNodeChangeset = Type.Object({
-	valueChange: Type.Optional(EncodedValueChange),
-	fieldChanges: Type.Optional(EncodedFieldChangeMap),
-	valueConstraint: Type.Optional(EncodedValueConstraint),
-	nodeExistsConstraint: Type.Optional(EncodedNodeExistsConstraint),
-});
-
-/**
- * Format for encoding as json.
- */
-type EncodedNodeChangeset = Static<typeof EncodedNodeChangeset>;
-
-const EncodedRevisionInfo = Type.Object({
-	revision: Type.Readonly(RevisionTagSchema),
-	rollbackOf: Type.ReadonlyOptional(RevisionTagSchema),
-});
-
-const EncodedModularChangeset = Type.Object({
-	maxId: Type.Optional(ChangesetLocalIdSchema),
-	changes: EncodedFieldChangeMap,
-	revisions: Type.ReadonlyOptional(Type.Array(EncodedRevisionInfo)),
-});
-
-type EncodedModularChangeset = Static<typeof EncodedModularChangeset>;
+import {
+	EncodedFieldChange,
+	EncodedFieldChangeMap,
+	EncodedModularChangeset,
+	EncodedNodeChangeset,
+} from "./modularChangeFormat";
 
 function makeV0Codec(
 	fieldKinds: ReadonlyMap<FieldKindIdentifier, FieldKind>,
